moved to audioPlayer package
