+++ common/src/main/java/com/allsoftdroid/common/base/store/audioPlayer/AudioPlayerEvent.kt
@@ -1,4 +1,4 @@
-package com.allsoftdroid.common.base.store
+package com.allsoftdroid.common.base.store.audioPlayer
 
 import com.allsoftdroid.common.base.extension.AudioPlayListItem
 import com.allsoftdroid.common.base.extension.AudioPlayerEventState
@@ -9,7 +9,7 @@ data class Next(val result:AudioPlayerEventState) : AudioPlayerEvent()
 data class Previous(val result: AudioPlayerEventState) : AudioPlayerEvent()
 data class Play(val result: AudioPlayerEventState) : AudioPlayerEvent()
 data class Pause(val result: AudioPlayerEventState) : AudioPlayerEvent()
-data class Initial(val default: AudioPlayerEventState):AudioPlayerEvent()
+data class EmptyEvent(val default: AudioPlayerEventState):AudioPlayerEvent()
 
 data class PlaySelectedTrack(val trackList : List<AudioPlayListItem>,val bookId:String,val bookName:String, val position:Int) : AudioPlayerEvent()
 data class TrackDetails(val trackTitle:String,val bookId: String, val position: Int):AudioPlayerEvent()
